messagesをcomponents/messages.vue に分離
