+++ pages/index.vue
@@ -1,39 +1,18 @@
 <template>
   <div>
-    <!--messages-->
-    <!--TODO:投稿がゼロ件でもloadingになってしまう-->
-    <div v-if="messages.length > 0">
-      <div v-for="message in messages">
-        {{ message.displayName }}:{{ message.content }}
-      </div>
-    </div>
-    <div v-else>
-      loading messages...
-    </div>
-    <!--writing area-->
+    <messages></messages>
     <post></post>
   </div>
 </template>
 
 <script>
-  import firebase from '~/plugins/firebase.js'
+  import messages from '~/components/messages.vue'
   import post from '~/components/post.vue'
 
-  const db = firebase.database();
-
   export default {
-    data () {
-      return {
-        messages: []
-      }
-    },
     components: {
+      messages,
       post
     },
-    mounted () {
-      db.ref('messages').on('child_added', snapshot => {
-        this.messages.push(snapshot.val());
-      })
-    }
   }
 </script>
